move the definition of `Parameters` to a separate module
